git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@2590 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/db/postgres/PGProcName.java
@@ -76,7 +76,7 @@ public class PGProcName
 		List<PGType> result = new ArrayList<PGType>(items.length);
 		for (String s : items)
 		{
-			int oid = Integer.valueOf(s.trim());
+			Long oid = Long.valueOf(s.trim());
 			PGType typ = typeMap.getTypeFromOID(oid);
 			if (typ != null)
 			{
